+++ react/CheckoutB2B.tsx
@@ -16,6 +16,7 @@ import {
 
 import 'vtex.country-codes/locales'
 
+import { CheckoutB2BProvider } from './CheckoutB2BContext'
 import { ContactInfos } from './components/ContactInfos'
 import {
   useClearCart,
@@ -101,7 +102,9 @@ function CheckoutB2BWrapper() {
   return (
     <QueryClientProvider client={queryClient}>
       <ToastProvider positioning="window">
-        <CheckoutB2BWithToast />
+        <CheckoutB2BProvider>
+          <CheckoutB2BWithToast />
+        </CheckoutB2BProvider>
       </ToastProvider>
     </QueryClientProvider>
   )