+++ CHANGELOG.md
@@ -9,7 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Add new column for payment address
+- Add new column for payment address.
+
+### Added
+
 - Pending state for order placed button.
 
 ## [0.0.8] - 2024-11-12